requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
python-dotenv==0.20.0
cloudinary==1.29.0